feat: Change plugin name to webhookreceiver, add app.kubernetes.io/name
Having hyphens in the name creates problems with Jinja2 templates.
Underscores are not allowed in Kubernetes service names.  To have one
name that works for both, and to follow the convention of other Tutor
plugins, we change the name to "webhookreceiver".

While we're at it, Tutor relies on the app.kubernetes.io/name label
(and label selectors) to be able to do things like "tutor k8s start
<servicename>". In the absence of those labels, this functionality
breaks. Add those where needed.

Reference:
https://kubernetes.io/docs/concepts/overview/working-with-objects/names/

Co-authored-by: Florian Haas <florian@citynetwork.eu>