fix: Single-quote all env values in k8s-deployments patch
Kubernetes wants all values passed in via the env dictionary to be
strings.

However, since MYSQL_PORT is an integer value, Kubernetes' YAML parser
reads it as such if unquoted, and stubbornly refuses to cast it to a
string.

Thus, quote that value, and so as not to run into similar issues in
the future, quote all env values in the k8s-deployments.