fix: Don't grant webhook_receiver_app_user superuser privileges
As documented in the webhook-receiver README, the webook receiver app
user (which the webhook receiver uses to make REST API calls to the
LMS) needs Staff privileges in order to be able to enroll users in all
courses — otherwise it would need to be added to all courses
individually as course staff, which isn't practical.

However, there is absolutely no need for superuser privileges. Remove
them from the LMS init hook.

Reference:
https://github.com/hastexo/webhook-receiver/#edx-oauth2-client